requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
attrs==24.2.0
av==12.3.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.6.1
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
einops==0.8.0
exceptiongroup==1.2.2
facexlib==0.3.0
filelock==3.14.0
filterpy==1.4.5
flake8==7.1.1
flatbuffers==24.3.25
ftfy==6.2.3
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.67.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
interrogate==1.7.0
isort==5.10.1
jmespath==0.10.0
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.7
llvmlite==0.41.1
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.2.2
mccabe==0.7.0
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv-full==1.5.0
-e git+https://github.com/hailsong/CLIP-IQA.git@5929d5b97cd8965f9acb9b0e1702b82f7ce179fb#egg=mmedit
model-index==0.1.11
mpmath==1.3.0
numba==0.58.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
Pillow==6.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
protobuf==5.28.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.0.2
torch==1.10.2
torchvision==0.11.3
tqdm==4.65.2
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.4
yapf==0.40.2
zipp==3.20.2
